karnak:rootdir:fix lock screen issues when sleeping
Over the period of time there was an issue with the device sleeping and whenit was waking up it tends to lag and system ui tends to crash so, i will have to remove this and let the kernel do its own memory management.

for 100 swappiness this has improved performance for the device.This tells the kernel to aggressively swap processes out of physical memory and move them to swap cache.

test:
cat proc/sys/vm/swappiness
returns: 100 for default

Change-Id: Ic2b721c7f55ed118a68a03d56855e4f74df13cba
Signed-off-by: Kai Jones <zaksenoj@icloud.com>