karnak:libaudio_shim: build from source
This fixes the issue #5 so how this happens is that when we tried to
plug in the headphones it seems to have the audio route coming out with
the headphones and the main speaker. the audiofix is a service  that
switches between headphones and speakers by controling
/sys/class/switch/h2w/state. the audio_shim library is a audio fix when
the headphones are disconnected and connected. the advantage of using
the audio_fix is that it tends to make the speaker loufder as Fire OS.
test:
build these libs and then ld_preload on init
and then played with audio. plugged in headphones to make sure the
sounds doesn't come out at the same time.

karnak:device.mk:remove libmtk_symbols package
forgot to added that change to the last commit

Change-Id: I32795841802a846f00915272906ca5ee8dfcb4bc
Signed-off-by: Kai Jones <zaksenoj@icloud.com>