karnak:move fstab and init files to vendor
The reason these files are moved to /vendor/etc/init/hw and
not /vendor/etc/init is because 'init' scans the /vendor/etc/init folder
for vendor service .rc fragments to import. That would either result
in duplicate imports or out-of-order imports of these hardware
specific .rc files and that *may* have unintended consequences.

for fstab.mt8163 I have added recoveryonly flag because system is
mounted by kernel now, and this flag will make
the fstab entry ignored during normal boot, and only used
for recovery.

In order to preserve the current import order for MTP, we move all
hardware specific .rc files to /vendor/etc/init/hw which is not
autoscanned by init BUT leave the init.${ro.hardware}.rc (i.e.
init.mt8163.rc in this case) in root for 'init' to find.
also symlink the fstab to / because the nvram depends on the fstab so it
can be copied to /data/nvram and the md32 binaries has to stay in the
rootfs because it won't work when it is in vendor.

Change-Id: I756d6074b0e85bfc5aafa6a8b1d5669eeb9488ab
Signed-off-by: Kai Jones <zaksenoj@icloud.com>