karnak:overlay:Remove dsp. prefix from audio and video power profiles
In the AOSP power_profile.xml file, the audio and video power profiles do not include the dsp. prefix. This commit removes the prefix from the power profiles in order to align with the AOSP power_profile.xml file. This change will ensure that the audio and video power profiles are properly recognized and used by the system.
Note: This commit only affects the audio and video power profiles, and does not modify any other aspect of the power_profile.xml file.

Change-Id: I453aa0a2054f28c3f580ab94123203fe83d02cd5
Signed-off-by: Kai Jones <kvjones0243@proton.me>