karnak:matrix:add device framework compatibility matrix file
This commit introduces a new file, the device framework compatibility
matrix, which outlines the compatibility between different devices and
the Android framework. The matrix is designed to help developers and
maintainers understand which devices are compatible with which versions
of the Android framework, and can be used to identify potential issues
when developing or testing on different devices.

The matrix is organized by device and lists the Android framework
versions that are compatible with each device. It also includes
information on any known issues or limitations that may affect
compatibility.

This change will help developers and maintainers understand the
compatibility of different devices with the Android framework, improving
the reliability and stability of the system.

Note: This change only affects the device framework compatibility matrix
and does not modify any other aspect of the system.
Change-Id: Ifed9a201ce7b15fa03bd0139f7c1c18695423596
Signed-off-by: Kai Jones <kvjones0243@proton.me>