karnak: camera: switch to Aperture
* LineageOS now has an awesome new camera app called Aperture! It is
based on Google’s (mostly) awesome CameraX library and provides a much
closer “to stock” camera app experience on many devices. Massive kudos
to developers @SebaUbuntu, @LuK1337, and @luca020400 who developed this
initially, designer Vazguard, and to the entire team for working to
integrate it into LineageOS and adapt it to our massive array of
supported devices!

* This has fixed the recording issues as well!

Change-Id: I7c4182e171ec4b9f4b2cffea868ad9fbcca7f8bd
Signed-off-by: Kai Jones <kvjones0243@proton.me>