karnak: libshims: Fix ABI issues
Avoid passing negative values for outBytesPerStride and
outBytesPerPixel as this can lead to crashes in some cases.
Use nullptr and allow the updated method to manage the situation.

Several constructor shims are implemented incorrectly because
they completely ignore the fact that constructors have 'this-call'
calling convention and should pass 'this ptr' as the first argument [1]
as well as its return value, at least according ARM ABI [2].

[1] https://review.lineageos.org/c/LineageOS/android_device_xiaomi_sdm710-common/+/360011/comments/eb6600df_3e2fb434
[2] https://github.com/ARM-software/abi-aa/blob/617079d8a0d45bec83d351974849483cf0cc66d5/cppabi32/cppabi32.rst?plain=1#L576

Co-authored-by: Ivan Vecera <ivan@cera.cz>

Change-Id: I7f4e6f240c5a0f69fadbeb8a4209dcb4a96d24ea
Signed-off-by: Kai Jones <19639817+488315@users.noreply.github.com>