karnak: fstab : Transition symlink from fstab to nvdata
As part of the transition to use `/vendor/nvdata` as a link to `/data/vendor/nvram`, this commit updates the `Android.mk` file. The previous `fstab.mt8163` symlink has been replaced with the `nvdata` symlink, ensuring that the new link is properly established in the build process. This change reflects the updated filesystem structure in the move from Fire OS 6 to Fire OS 7. The `ALL_DEFAULT_INSTALLED_MODULES` now includes the `nvdata` symlink.